+++ inferia/core/config.py
@@ -4,7 +4,7 @@ from typing import List, Optional
 import yaml
 from pydantic import BaseModel
 
-from cogito.core.exceptions import ConfigFileNotFoundError
+from inferia.core.exceptions import ConfigFileNotFoundError
 
 class ArgConfig(BaseModel):
     name: str
@@ -64,7 +64,7 @@ class ServerConfig(BaseModel):
     @classmethod
     def default(cls):
         return cls(
-            name='Cogito ergo infero',
+            name='Sapientia per Inferentiam',
             description='Inference server',
             version='0.1.0',
             fastapi=FastAPIConfig.default(),
@@ -83,9 +83,9 @@ class TrainingConfig(BaseModel):
         return cls()
 
 
-class CogitoConfig(BaseModel):
+class InferiaConfig(BaseModel):
     """
-    Cogito configuration.
+    Inferia configuration.
     """
     server: ServerConfig
     training: TrainingConfig
@@ -99,11 +99,11 @@ class ConfigFile(BaseModel):
     """
     Configuration file.
     """
-    cogito: CogitoConfig
+    inferia: InferiaConfig
 
     @classmethod
     def default(cls):
-        return cls(cogito=CogitoConfig.default())
+        return cls(inferia=InferiaConfig.default())
 
     @classmethod
     def exists(cls, file_path: str) -> bool: